refactor(tests): split configuration tests to new file
