+++ tests/api_test.py
@@ -12,112 +12,9 @@ import coverage
 import mock
 import pytest
 import sh
-try:
-    import yaml
-except ImportError:
-    yaml = None
 
 from coveralls import Coveralls
 from coveralls.api import log
-
-
-@mock.patch.object(Coveralls, 'config_filename', '.coveralls.mock')
-class Configration(unittest.TestCase):
-
-    def setUp(self):
-        with open('.coveralls.mock', 'w+') as fp:
-            fp.write('repo_token: xxx\n')
-            fp.write('service_name: jenkins\n')
-
-    def tearDown(self):
-        os.remove('.coveralls.mock')
-
-    @pytest.mark.skipif(yaml is None, reason='requires pyyaml')
-    @mock.patch.dict(os.environ, {}, clear=True)
-    def test_local_with_config(self):
-        cover = Coveralls()
-        assert cover.config['service_name'] == 'jenkins'
-        assert cover.config['repo_token'] == 'xxx'
-        assert 'service_job_id' not in cover.config
-
-    def generate_import_mock(self, bad_module_name, msg):
-        """ Returns replacement for builting import function, which raises
-        exception on importing appropriate module"""
-
-        origin = __import__
-        def import_func(name, *args):
-            if name == bad_module_name:
-                raise ImportError(msg)
-            else:
-                return origin(name, *args)
-        return import_func
-
-    @mock.patch.object(log, 'warning')
-    def test_local_with_config_without_yaml_module(self, mock_logger):
-        """test local with config in yaml, but without yaml-installed"""
-
-        if sys.version_info < (3, 0):
-            builtin_import_func = '__builtin__.__import__'
-        else:
-            builtin_import_func = 'builtins.__import__'
-
-        yaml_import_mock = self.generate_import_mock('yaml', 'No module named yaml')
-        try:
-            # patching of `import` function of the Coveralls module (it shoud throw ImportException):
-            with mock.patch(builtin_import_func, side_effect=yaml_import_mock):
-                Coveralls()
-        except Exception:
-            pass
-        mock_logger.assert_called_once_with('Seems, like some modules are not installed: %s', mock.ANY)
-
-@mock.patch.object(Coveralls, 'config_filename', '.coveralls.mock')
-class NoConfig(unittest.TestCase):
-
-    @mock.patch.dict(os.environ, {'TRAVIS': 'True', 'TRAVIS_JOB_ID': '777'}, clear=True)
-    def test_travis_no_config(self):
-        cover = Coveralls()
-        assert cover.config['service_name'] == 'travis-ci'
-        assert cover.config['service_job_id'] == '777'
-        assert 'repo_token' not in cover.config
-
-    @mock.patch.dict(os.environ, {'TRAVIS': 'True', 'TRAVIS_JOB_ID': '777', 'COVERALLS_REPO_TOKEN': 'yyy'}, clear=True)
-    def test_repo_token_from_env(self):
-        cover = Coveralls()
-        assert cover.config['service_name'] == 'travis-ci'
-        assert cover.config['service_job_id'] == '777'
-        assert cover.config['repo_token'] == 'yyy'
-
-    @mock.patch.dict(os.environ, {}, clear=True)
-    def test_misconfigured(self):
-        with pytest.raises(Exception) as excinfo:
-            Coveralls()
-        assert str(excinfo.value) == 'Not on Travis or CircleCI. You have to provide either repo_token in .coveralls.mock or ' \
-                                     'set the COVERALLS_REPO_TOKEN env var.'
-
-    @mock.patch.dict(os.environ, {'CIRCLECI': 'True',
-                                  'CIRCLE_BUILD_NUM': '888',
-                                  'CI_PULL_REQUEST': 'https://github.com/org/repo/pull/9999'}, clear=True)
-    def test_circleci_no_config(self):
-        cover = Coveralls()
-        assert cover.config['service_name'] == 'circle-ci'
-        assert cover.config['service_job_id'] == '888'
-        assert cover.config['service_pull_request'] == '9999'
-
-    @mock.patch.dict(os.environ, {'APPVEYOR': 'True',
-                                  'APPVEYOR_BUILD_ID': '1234567',
-                                  'APPVEYOR_PULL_REQUEST_NUMBER': '1234'}, clear=True)
-    def test_appveyor_no_config(self):
-        cover = Coveralls(repo_token='xxx')
-        assert cover.config['service_name'] == 'appveyor'
-        assert cover.config['service_job_id'] == '1234567'
-        assert cover.config['service_pull_request'] == '1234'
-
-    @mock.patch.dict(os.environ, {'BUILDKITE': 'True',
-                                  'BUILDKITE_JOB_ID': '1234567'}, clear=True)
-    def test_buildkite_no_config(self):
-        cover = Coveralls(repo_token='xxx')
-        assert cover.config['service_name'] == 'buildkite'
-        assert cover.config['service_job_id'] == '1234567'
 
 
 def assert_coverage(actual, expected):
